increasing activity

+++ src/main/java/com/rickmorty/DTO/CharacterDto.java
@@ -2,12 +2,16 @@ package com.rickmorty.DTO;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.util.ArrayList;
 import java.util.List;
 import com.rickmorty.DTO.CharacterDto;
+import com.rickmorty.enums.Gender;
+import com.rickmorty.enums.LifeStatus;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record CharacterDto(
-        @JsonProperty("id") int id,
+        @JsonProperty("id") Long id,
         @JsonProperty("name") String name,
         @JsonProperty("status") String status,
         @JsonProperty("species") String species,
@@ -17,4 +21,6 @@ public record CharacterDto(
         @JsonProperty("episode") List<String> episode,
         @JsonProperty("location") LocationCharacterDto location
 
-) {}
+) {
+
+}
